# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Flora Bouquets API",
    "version": "1.0.0",
    "description": "REST API for managing the Flora bouquet collection."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "tags": [
    {
      "name": "Bouquets",
      "description": "Bouquet collection operations"
    }
  ],
  "paths": {
    "/api/bouquets": {
      "get": {
        "tags": [
          "Bouquets"
        ],
        "summary": "Get all bouquets",
        "operationId": "getBouquets",
        "responses": {
          "200": {
            "description": "Bouquet collection",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Bouquet"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Bouquets"
        ],
        "summary": "Create a bouquet",
        "description": "Creates a bouquet. The photoURL is generated by the server.",
        "operationId": "createBouquet",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BouquetCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bouquet created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bouquet"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/bouquets/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/BouquetId"
        }
      ],
      "get": {
        "tags": [
          "Bouquets"
        ],
        "summary": "Get a bouquet by id",
        "operationId": "getBouquetById",
        "responses": {
          "200": {
            "description": "Bouquet found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bouquet"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": [
          "Bouquets"
        ],
        "summary": "Update a bouquet",
        "description": "Updates one or more provided bouquet fields. An empty request body is invalid.",
        "operationId": "updateBouquet",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BouquetUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bouquet updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bouquet"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "Bouquets"
        ],
        "summary": "Delete a bouquet",
        "operationId": "deleteBouquet",
        "responses": {
          "200": {
            "description": "Bouquet deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/bouquets/{id}/favorite": {
      "patch": {
        "tags": [
          "Bouquets"
        ],
        "summary": "Update bouquet favorite status",
        "operationId": "updateBouquetFavorite",
        "parameters": [
          {
            "$ref": "#/components/parameters/BouquetId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Favorite status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bouquet"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/bouquets/{id}/photo": {
      "patch": {
        "tags": [
          "Bouquets"
        ],
        "summary": "Upload a new bouquet photo",
        "description": "Accepts one JPEG, PNG, or WebP image in the photo multipart field.",
        "operationId": "updateBouquetPhoto",
        "parameters": [
          {
            "$ref": "#/components/parameters/BouquetId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "photo"
                ],
                "properties": {
                  "photo": {
                    "type": "string",
                    "format": "binary",
                    "description": "JPEG, PNG, or WebP image. Maximum size: 5 MB."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bouquet photo updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoUploadResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "BouquetId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Positive integer bouquet identifier",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "example": 1
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request data",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "message": "price must be greater than or equal to 0"
            }
          }
        }
      },
      "NotFound": {
        "description": "Bouquet was not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "message": "Not found"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Unexpected server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "message": "Internal server error"
            }
          }
        }
      }
    },
    "schemas": {
      "Bouquet": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "title",
          "description",
          "price",
          "photoURL",
          "favorite"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1,
            "readOnly": true,
            "example": 1
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "example": "Midnight Garden"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "example": "A spellbinding bouquet of indigo flowers and eucalyptus."
          },
          "price": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "maximum": 99999999.99,
            "example": 98
          },
          "photoURL": {
            "type": "string",
            "format": "uri",
            "readOnly": true,
            "example": "https://s.gravatar.com/avatar/example"
          },
          "favorite": {
            "type": "boolean",
            "default": false,
            "example": false
          }
        }
      },
      "BouquetCreate": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "title",
          "description",
          "price"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "example": "Spring Garden"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "example": "A fresh arrangement of seasonal flowers."
          },
          "price": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "maximum": 99999999.99,
            "example": 75.5
          },
          "favorite": {
            "type": "boolean",
            "default": false,
            "example": false
          }
        }
      },
      "BouquetUpdate": {
        "type": "object",
        "additionalProperties": false,
        "minProperties": 1,
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "example": "Updated Spring Garden"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "example": "An updated seasonal flower arrangement."
          },
          "price": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "maximum": 99999999.99,
            "example": 85
          },
          "favorite": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "FavoriteUpdate": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "favorite"
        ],
        "properties": {
          "favorite": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "PhotoUploadResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "photoURL"
        ],
        "properties": {
          "photoURL": {
            "type": "string",
            "format": "uri",
            "example": "http://localhost:3000/photos/550e8400-e29b-41d4-a716-446655440000.jpg"
          }
        }
      },
      "DeleteResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "Bouquet deleted"
          }
        }
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "Not found"
          }
        }
      }
    }
  }
}